Payment done

+++ PaymentsService/Payments.Domain/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -29,8 +29,8 @@ namespace Payments.Domain.Migrations
 
                     NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
 
-                    b.Property<int>("Price")
-                        .HasColumnType("integer");
+                    b.Property<decimal>("Price")
+                        .HasColumnType("numeric");
 
                     b.Property<int>("ReferenceId")
                         .HasColumnType("integer");
